Makefile: Polly: Remove -polly-opt-fusion=max cmdline option
The command line option -polly-opt-fusion has been completely removed [1] in the latest LLVM toolchain (and in the latest Clang 14 build)

The cmdline option was removed previously [2], though the main switch remained because of a human error
which kept this option functioning in the polly optimizer. But in reality, it did nothing since then.
Now, it has been completely removed. We should also follow that to use the latest LLVM toolchain without
any compilation error.

The name of the option is misleading and has been renamed by isl to
"serialize-sccs". Instead of also renaming the option, it has been removed by llvm.

The option is still accessible using the following cmdline option:

[ -polly-isl-arg=--no-schedule-serialize-sccs ]

[1] https://github.com/llvm/llvm-project/commit/cb879d00d8d54383447c1d94ec4ef0b28873dcaa

[2] https://github.com/llvm/llvm-project/commit/07e7cb9433daa80e7a8b418cb334c663abe68202

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Devang <devangsingh369@gmail.com>
Signed-off-by: firebird11 <hbgassel@gmail.com>