[SQUASH] touchscreen: Optimize interrupt CPU usage & Adapt PM QoS usage to incoherent OnePlus code
* Adapt from kerneltoast/android_kernel_google_floral@5726443
* This also reduces input latency
Since OnePlus frees the IRQ, we need to remove the PM QoS request prior
to that in order to unregister the IRQ affinity notifier in the PM QoS
request. We also need to move the pm_qos_add_request() operation to
come before the IRQ is registered, to prevent the IRQ handler from using
the PM QoS structures while they're uninitialized.